moving tensorboard functions
